% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{calculateMI}
\alias{calculateMI}
\title{Calculate mutual information.}
\usage{
calculateMI(x, y, num_bins = 10, num_permutations = 0)
}
\arguments{
\item{x}{A random variable (vector/factor).}

\item{y}{Another random variable (vector/factor).}

\item{num_bins}{The number of intervals into which x and y is to be cut.}

\item{num_permutations}{Number of permutations to use, if greater than 0, compute empirical p-values using a permutation test.}
}
\value{
A list include MI and pvalue (if num_permutations > 0).
}
\description{
Calculate mutual information.
}
