% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{calculateSpatialMI}
\alias{calculateSpatialMI}
\title{Calculate spatial mutual information (z to x,y).}
\usage{
calculateSpatialMI(x, y, z, num_bins = 10, num_permutations = 0)
}
\arguments{
\item{x}{A random variable (vector/factor) on the first axis.}

\item{y}{A random variable (vector/factor) on the second axis.}

\item{z}{The random variable (vector/factor) to be calculate to x and y.}

\item{num_bins}{The number of intervals into which x, y and z is to be cut.}

\item{num_permutations}{Number of permutations to use, if greater than 0, compute empirical p-values using a permutation test.}
}
\value{
A list include MI and pvalue (if num_permutations > 0).
}
\description{
Calculate spatial mutual information (z to x,y).
}
