% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{doBinarization}
\alias{doBinarization}
\title{Title Binarization the pathway activity score by Kmeans cluster.}
\usage{
doBinarization(score.data, n.cluster = 2, method = "GMM")
}
\arguments{
\item{score.data}{A data frame store cell pathway activity score in column 1.}

\item{n.cluster}{Number of clusters, 2 or 3 is recommended, default set to 2.}

\item{method}{Binarization method, The choices are "GMM", "kmeans", default set to "GMM".
"GMM" is Gaussian Mixture Model, "kmeans" is K-means cluster.}
}
\value{
A data frame store the binarization label.
}
\description{
Title Binarization the pathway activity score by Kmeans cluster.
}
