% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{computeScore}
\alias{computeScore}
\title{Compute the pathway activity score.}
\usage{
computeScore(
  object,
  regression.data = NULL,
  pvalue = 0.05,
  weight = FALSE,
  normalize = "z-score"
)
}
\arguments{
\item{object}{A Seurat object include MCA data in reductions slot.}

\item{regression.data}{Regression data frame, generated by doRegression.}

\item{pvalue}{The cutoff p value of linear regression result, default set to 0.05.}

\item{weight}{Whether calculate the weighted activity score, weighted means that the variance contribution rate of MCA reduction is used as the weight of each dimension. Default set to FALSE.}

\item{normalize}{Normalize method, The choices are NULL, "z-score" and "bg-score". "z-socre" means z-score normalize, "bg-score" means normalize by background activity score. Default set to "z-score".}
}
\value{
A data frame store single-cell pathways activity score.
}
\description{
Compute the pathway activity score.
}
