% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{computeScore}
\alias{computeScore}
\title{Step4: Compute the pathway activity score.}
\usage{
computeScore(object, regression.data = NULL, pvalue = 0.05, weight = FALSE)
}
\arguments{
\item{object}{A Seurat object include MCA data in reductions slot.}

\item{regression.data}{Regression data frame, generated by doRegression.}

\item{pvalue}{The cutoff p value of linear regression result.}

\item{weight}{Whether calculate the weighted activity score, weighted means that the variance contribution rate of MCA reduction is used as the weight of each dimension.}
}
\value{
A data frame store cell pathways activity score.
}
\description{
Step4: Compute the pathway activity score.
}
