% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{computeScore}
\alias{computeScore}
\title{Step4: Compute the pathway activity score.}
\usage{
computeScore(object, regression.data = NULL)
}
\arguments{
\item{object}{A Seurat object include MCA data in reductions slot.}

\item{regression.data}{Regression data frame, generated by doRegression.}
}
\value{
A data frame store cell pathway activity score.
}
\description{
Step4: Compute the pathway activity score.
}
