% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{doRegression}
\alias{doRegression}
\title{Step3: Do regression of gene loading (Y) and gene rate (X)}
\usage{
doRegression(object, gene.rate = NULL)
}
\arguments{
\item{object}{A Seurat object include MCA data in reductions slot.}

\item{gene.rate}{Gene rate data frame, generated by getGeneRate.}
}
\value{
A data frame include t and p value of pathways and corresponding background, dim=(4*num(pathways), nmcs).
}
\description{
Step3: Do regression of gene loading (Y) and gene rate (X)
}
