% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{getGeneRate}
\alias{getGeneRate}
\title{Step2: Get gene occurrence rate in background and signature genesets}
\usage{
getGeneRate(background.file = NULL, pathway.file = NULL, pathway.name = NULL)
}
\arguments{
\item{background.file}{Background genesets file, in GMT format.}

\item{pathway.file}{Pathway geneset file, in txt fotmat, one gene in one line.}

\item{pathway.name}{Pathway geneset name.}
}
\value{
A data frame include 2 col, colnames is c("background", pathway.name).
}
\description{
Step2: Get gene occurrence rate in background and signature genesets
}
