% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaaSc.R
\name{getGeneRate}
\alias{getGeneRate}
\title{Step2: Get gene occurrence rate in background and pathway genesets}
\usage{
getGeneRate(background.file = NULL, pathway.geneset = NULL)
}
\arguments{
\item{background.file}{Background genesets file, in GMT format.}

\item{pathway.geneset}{Pathway geneset, a R object of list which include tags (pathway name) and values (gene list).}
}
\value{
A data frame include background and pathways gene rate.
}
\description{
Step2: Get gene occurrence rate in background and pathway genesets
}
